feat(suite-native): authorize passphrased device
